chore(cli): Put status code in status module, setup in setup
