+++ src/main.rs
@@ -15,7 +15,7 @@ fn main() -> Result<(), Box<dyn error::Error>> {
     fontship::show_welcome();
     match args.subcommand {
         Subcommand::Make { target } => make::run(target),
-        Subcommand::Setup {} => setup::run(),
-        Subcommand::Status { path } => status::run(path),
+        Subcommand::Setup { path } => setup::run(path),
+        Subcommand::Status {} => status::run(),
     }
 }
